Factor out all but top level functions in separate file
